// src/payload-types.ts
export interface User {
	id: string;
	name?: string | null;
	roles?: ("super-admin" | "admin" | "editor")[] | null;
	updatedAt: string;
	createdAt: string;
	email: string;
	resetPasswordToken?: string | null;
	resetPasswordExpiration?: string | null;
	salt?: string | null;
	hash?: string | null;
	loginAttempts?: number | null;
	lockUntil?: string | null;
	sessions?:
		| {
				id: string;
				createdAt?: string | null;
				expiresAt: string;
		  }[]
		| null;
	password?: string | null;
}